+++ frontend/style.css
@@ -192,8 +192,14 @@ html, body {
   bottom: 15px;
   left: 15px;
 
-  height: 135px;
-  width: 135px;
+  height: 8em;
+  width: 8em;
+}
+
+/* let clicks pass through the axes helper */
+#axes-helper-frame,
+#axes-helper-frame * {
+  pointer-events: none;
 }
 
 #controls {
